# openapi.yaml
openapi: 3.0.3
info:
  title: Media Analyzer API
  description: |
    Comprehensive API for analyzing social media content, particularly Instagram Reels, with advanced capabilities including:
    - **Content Analysis**: Full compliance checking and brand safety analysis
    - **Keyword Extraction**: Advanced semantic keyword extraction with sentiment analysis
    - **Media Processing**: Audio transcription (ASR) and optical character recognition (OCR)
    - **Named Entity Recognition**: ML-based entity extraction and semantic similarity
    - **Multilingual Support**: English, Hindi, and Hinglish content processing
    
    ## System Architecture
    
    The system consists of three main components:
    
    1. **Main API Server (Node.js/Fastify)** - Port 3000
    2. **Python Worker (FastAPI)** - Port 8000
    3. **Database & Queue System** - SQLite with Redis
    
    ## Authentication
    
    All protected endpoints require JWT authentication via Bearer token in the Authorization header.
    
  version: 1.0.0
  contact:
    name: Media Analyzer API Support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Main API Server (Development)
  - url: http://localhost:8000
    description: Python Worker Server (Development)
  - url: https://api.media-analyzer.com
    description: Production Server
...
paths:
...
  /v1/keywords/extract:
    post:
      tags:
        - Keywords
      summary: Extract keywords (Basic)
      description: |
        Extracts searchable keywords from Instagram Reels for content discovery.
        
        **Features:**
        - Primary and secondary keywords
        - Hashtags and mentions
        - Topic categorization
        - Searchable terms optimization
      operationId: extractKeywords
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KeywordExtractionRequest'
            examples:
              instagram_basic:
                summary: Instagram Reel extraction
                value:
                  instagramReelUrl: https://www.instagram.com/reel/ABC123DEF456/
                  languageHint: en
              youtube_shorts:
                summary: YouTube Shorts extraction
                value:
                  shortVideoUrl: https://www.youtube.com/shorts/ABC123DEF456
                  languageHint: en
              with_cookies:
                summary: With Chrome cookies
                value:
                  instagramReelUrl: https://www.instagram.com/reel/ABC123DEF456/
                  languageHint: en
                  cookieOptions:
                    browserCookies: chrome
      responses:
        '200':
          description: Keywords extracted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KeywordExtractionResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/keywords/extract-enhanced:
    post:
      tags:
        - Keywords
      summary: Extract keywords (Enhanced)
      description: |
        Enhanced keyword extraction with advanced semantic analysis.
        
        **Additional Features:**
        - N-gram extraction and phrase analysis
        - Sentiment analysis with emotions
        - Intent detection
        - Entity extraction (brands, products, people, etc.)
        - Multilingual support with auto-detection
      operationId: extractKeywordsEnhanced
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnhancedKeywordExtractionRequest'
            examples:
              instagram_full_features:
                summary: Instagram Reel with full enhanced features
                value:
                  instagramReelUrl: https://www.instagram.com/reel/ABC123DEF456/
                  languageHint: en
                  options:
                    includeNgrams: true
                    includeSentiment: true
                    includeIntent: true
                    includeEntities: true
                  cookieOptions:
                    browserCookies: chrome
              youtube_shorts_enhanced:
                summary: YouTube Shorts with enhanced features
                value:
                  shortVideoUrl: https://www.youtube.com/shorts/ABC123DEF456
                  languageHint: en
                  options:
                    includeNgrams: true
                    includeSentiment: true
                    includeIntent: true
                    includeEntities: true
      responses:
        '200':
          description: Enhanced keywords extracted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnhancedKeywordExtractionResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login or /auth/demo-token
  schemas:
...
    KeywordExtractionRequest:
      type: object
      oneOf:
        - required: [instagramReelUrl]
        - required: [shortVideoUrl]
      properties:
        instagramReelUrl:
          type: string
          format: uri
          pattern: '^https?://(www\.)?instagram\.com/(reel|reels|p)/[A-Za-z0-9_-]+/?$'
          example: https://www.instagram.com/reel/ABC123DEF456/
        shortVideoUrl:
          type: string
          format: uri
          pattern: '^https?://((www\.)?instagram\.com/(reel|reels|p)/[A-Za-z0-9_-]+/?|(www\.)?youtube\.com/shorts/[A-Za-z0-9_-]+|youtu\.be/[A-Za-z0-9_-]+)$'
          example: https://www.youtube.com/shorts/ABC123DEF456
          description: "Instagram Reel or YouTube Shorts URL"
        languageHint:
          type: string
          example: en
        cookieOptions:
          $ref: '#/components/schemas/CookieOptions'
    EnhancedKeywordExtractionRequest:
      allOf:
        - $ref: '#/components/schemas/KeywordExtractionRequest'
        - type: object
          properties:
            options:
              type: object
              properties:
                includeNgrams:
                  type: boolean
                  default: true
                includeSentiment:
                  type: boolean
                  default: true
                includeIntent:
                  type: boolean
                  default: true
                includeEntities:
                  type: boolean
                  default: true
            async:
              type: boolean
              default: false
    KeywordExtractionResponse:
      type: object
      properties:
        keywords:
          type: object
          properties:
            primary:
              type: array
              items:
                type: string
              example: ["fashion", "style", "outfit", "trendy"]
            secondary:
              type: array
              items:
                type: string
              example: ["clothes", "accessories", "shopping", "beauty"]
            hashtags:
              type: array
              items:
                type: string
              example: ["#fashion", "#style", "#ootd", "#trending"]
            mentions:
              type: array
              items:
                type: string
              example: ["@brand", "@influencer"]
            topics:
              type: array
              items:
                type: string
              example: ["fashion", "lifestyle"]
        metadata:
          type: object
          properties:
            caption:
              type: string
            transcript:
              type: string
            ocrText:
              type: string
            duration:
              type: number
            username:
              type: string
        searchableTerms:
          type: array
          items:
            type: string
        timings:
          $ref: '#/components/schemas/Timings'
    EnhancedKeywordExtractionResponse:
      allOf:
        - $ref: '#/components/schemas/KeywordExtractionResponse'
        - type: object
          properties:
            keywords:
              type: object
              properties:
                phrases:
                  type: array
                  items:
                    type: object
                    properties:
                      text:
                        type: string
                      frequency:
                        type: integer
                      significance:
                        type: number
            topics:
              type: object
              properties:
                primary:
                  type: object
                  properties:
                    category:
                      type: string
                    subcategory:
                      type: string
                    confidence:
                      type: number
                secondary:
                  type: array
                  items:
                    type: object
                    properties:
                      category:
                        type: string
                      confidence:
                        type: number
            sentiment:
              type: object
              properties:
                overall:
                  type: string
                  enum: [positive, neutral, negative]
                score:
                  type: number
                  minimum: -5
                  maximum: 5
                comparative:
                  type: number
                emotions:
                  type: array
                  items:
                    type: string
            intent:
              type: object
              properties:
                primary:
                  type: string
                  enum: [educate, entertain, promote, inspire, inform]
                secondary:
                  type: array
                  items:
                    type: string
                confidence:
                  type: number
            entities:
              type: object
              properties:
                brands:
                  type: array
                  items:
                    type: string
                products:
                  type: array
                  items:
                    type: string
                people:
                  type: array
                  items:
                    type: string
                prices:
                  type: array
                  items:
                    type: string
                locations:
                  type: array
                  items:
                    type: string

    # ==================== Python Worker Schemas ====================
...
    CookieOptions:
      type: object
      properties:
        browserCookies:
          type: string
          enum: [chrome, firefox, safari, edge, opera, brave]
        cookiesFile:
          type: string
    Timings:
      type: object
      properties:
        totalMs:
          type: integer
          description: Total processing time in milliseconds
        stages:
          type: object
          additionalProperties:
            type: integer
          example:
            extract: 800
            asr: 1200
            ocr: 300
            processing: 200
...
  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: string
                example: VALIDATION_ERROR
              message:
                type: string
                example: Invalid request data
              details:
                type: array
                items:
                  type: object
                  properties:
                    path:
                      type: array
                      items:
                        type: string
                    message:
                      type: string
    UnauthorizedError:
      description: Unauthorized
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: string
                example: UNAUTHORIZED
              message:
                type: string
                example: Invalid or missing authentication token
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: string
                example: INTERNAL_ERROR
              message:
                type: string
                example: An internal error occurred
              details:
                type: string